requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-restframework==0.0.1
djangorestframework==3.13.1
idna==3.3
python-decouple==3.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
stripe==2.65.0
urllib3==1.26.8
